Butterfly/butterflyApp/producers/gitlab-producer/requirements.txt
Flask==1.0.3
kafka-python==1.4.6